% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmnwoo.R
\name{pmnwoo}
\alias{pmnwoo}
\title{Title}
\usage{
pmnwoo(counties, townships)
}
\arguments{
\item{counties}{this is an analysis at county level}

\item{townships}{analysis at township level}
}
\value{
one dataframe
}
\description{
Title
}
\examples{
x<-1

}
